# src/fxn/dto/cancel-subscription-request.dto.ts
import { IsNumber, IsOptional } from 'class-validator';

import { IsSolanaPublicKey } from '../../utils/custom-validators';

export default class CancelSubscriptionRequestDto {
  @IsSolanaPublicKey()
  dataProvider: string;

  @IsNumber()
  qualityScore: number;

  @IsOptional()
  @IsSolanaPublicKey()
  nftTokenAccount?: string;
}